test: hoist shared reset-server FFI helper into test_support
